django/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
mysqlclient==2.2.7
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.3
